Don't filter equivalent Episode when its Publisher has no strong Brands
Previously if a candidate Episode's Brand wasn't a strong equivalent of the
subject's Brand that equivalence was undone. Now, if the subject's Brand
has no strong equivalences from the candidate's publisher the
equivalence is left in tact. That is, the equivalence is only undone
when the subject's Brand has a non-empty set of strong equivalences of the same publisher
as the candidate's Brand and that set does not contain the candidate's Brand.