Magpie ingest job is restartable, will only ingest changed files
The Magpie ingest job keeps persistent track of the lastModifiedTime of
the last results file successfully ingested. File are ingested by order
of lastModifiedTime. The job will ingest all files that have a
lastModifiedTime greater (after) the persisted time. This means the job
is restartable though it will re-ingest in its entirety the job that was
in progress when the previous run was stopped.