Compute equivalence for fetched Content, Facebook-specific equiv updater
When new Content is fetched through the UriFetchingQueryExecutor run it
through the ContentEquivalenceUpdater if it is of a specified set of
Publishers.

The EquivModule's scheduled tasks are separated into a new
EquivTaskModule.

The Facebook equivalence updater relies on title searching and resolving
aliases. The filter requiring an Item score is removed along with the
Publisher filter allowing Facebook content to be matched to Facebook
content.