Add 'strict' mode to equivalence Episode filtering
Configure Episode equivalence filtering with a strict mode flag. In
strict mode an Item's candidate equivalent's Container must be
equivalent to the Item's Container. That is, for a candiate equivalent,
E, of item, I, E's Container, Ec, must be equivalent to I's Container,
Ic. In relaxed mode, Ec need not be equivalent to Ic but there must be
no other equivalent container for that source.

Currently all filters are in relaxed mode