Equiv sequence scoring null for top-level series parent mismatch
When scoring two episodes by sequence either both must have a top-level
series as parent or neither, otherwise the score is null. Episodes in a
standard Brand/Series/Episode hierarchy will score null against an
Episode in a Series/Episode hierarchy.

Detection of Episodes as children of top-level Series is through
checking the episode's parent ref is the same as its series ref.