Ingest PA repeat flag for all channels
Simplifies the interpretation of the Broadcast _repeat_ flag for PA
data. A PA 'yes' flag is ingested as a repeat for __all__ channels. A PA
'no' flag or a missing flag is interpreted as `null`. The rational is a
'yes' flag is _definitely_ a repeat, anything else we can't be sure.