Fix error getting absent person
The PersonQueryResolver returns Optional.absent() when a Person's
source is disabled for the request's application. The subsequent
Optional.get() therefore throws an ISE. This avoids the get() by
simply testing isPresent(). It can't distinguish between a non-
existent person and an invisible person at this point so the return
code is 404.