Improve ScheduleController error message when key and source missing
It's an error not to provide a publisher or API key to /schedule but
resulting message was "Need precedence enabled" didn't described that
very well. If neither param is supplied then "You must supply a
publisher or API key" is displayed.

If an API key without precedence is supplied without a publishers then
the original message is displayed.