Output parent channel on items
Parent channels are generally useful for branding
purposes. Therefore we'll output if channel information
is requested with the CHANNEL_SUMMARY annotation.

We are also overriding the images provided for parent
channels because we don't yet receive images on parent
channels (a.k.a. stations) from PA, currently our only
source of channels.