Relax constraint on number of episodes in brand
We generally don't want picks to include items from brands
with many episodes, since they're generally not that
interesting. However, this systematically excludes some
genres, notably soaps. We'll therefore not apply the
exclusion based on number of episodes if an item has a
particular genre. Initially, we're setting that to just
soaps, but this list may be expanded.